v1-oldapi/requirements.txt
pandas==1.3.2
numpy==1.21.0
pytest==5.1.2
fake_useragent==0.1.11
responses==0.10.6
matplotlib==3.3.3
youtube_search==0.1.3
pytz==2021.1
seaborn==0.11.2
requests==2.26.0
tweepy==3.10.0
twython==3.8.2
lxml==4.6.3
beautifulsoup4==4.10.0
Pillow==8.3.2
python-dateutil==2.8.2
PyYAML==5.4.1
